requirements.txt
aiogram==3.6.0
asyncpg==0.29.0
apscheduler==3.10.4
python-dotenv==1.0.1
google-genai==0.2.2
pydantic==2.7.1
setuptools>=70.0.0
wheel>=0.43.0
aiohttp==3.9.3
